Split component state mapping to containers
