+++ src/components/Seat.js
@@ -1,16 +1,6 @@
 import React from "react";
-import { connect } from "react-redux";
 import "../App.css";
-import { takeSeatAction } from "../actions/takeSeatAction";
 import { func, string } from "prop-types";
-
-const mapStateToProps = state => ({
-    ...state,
-});
-
-const mapDispatchToProps = dispatch => ({
-    takeSeatAction: id => dispatch(takeSeatAction(id)),
-});
 
 const propTypes = {
     takeSeatAction: func.isRequired,
@@ -21,7 +11,7 @@ const Seat = props => {
     const seatLocation = props.seatLocation;
 
     const takeSeat = event => { // eslint-disable-line
-    // eslint-disable-line
+        // eslint-disable-line
         props.takeSeatAction(seatLocation);
     };
 
@@ -30,4 +20,4 @@ const Seat = props => {
 
 Seat.propTypes = propTypes;
 
-export default connect(mapStateToProps, mapDispatchToProps)(Seat);
+export default Seat;
