+++ src/containers/Seat.js
@@ -1,0 +1,11 @@
+import { connect } from "react-redux";
+import { takeSeatAction } from "../actions/takeSeatAction";
+import Seat from '../components/Seat';
+
+const mapStateToProps = state => ({}); //eslint-disable-line
+
+const mapDispatchToProps = dispatch => ({
+    takeSeatAction: id => dispatch(takeSeatAction(id)),
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(Seat);
