+++ src/containers/Sitsit.js
@@ -1,0 +1,8 @@
+import { connect } from "react-redux";
+import Sitsit from '../components/Sitsit';
+
+const mapStateToProps = state => ({}); //eslint-disable-line
+
+const mapDispatchToProps = state => ({}); //eslint-disable-line
+
+export default connect(mapStateToProps, mapDispatchToProps)(Sitsit);
