+++ src/containers/Table.js
@@ -1,0 +1,8 @@
+import { connect } from "react-redux";
+import Table from "../components/Table";
+
+const mapStateToProps = state => ({}); // eslint-disable-line
+
+const mapDispatchToProps = dispatch => ({}); // eslint-disable-line
+
+export default connect(mapStateToProps, mapDispatchToProps)(Table);
